pyproject.toml
[tool.poetry]
name = "vkquick_bot"
version = "0.1.0"
description = ""
authors = ["deknowny <deknowny@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
python-dotenv = "^0.17.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
